+++ packages/terra-infinite-list/src/terra-dev-site/doc/example/ProgressiveLoadingExample.jsx
@@ -2,10 +2,14 @@ import React from 'react';
 /* eslint-disable import/no-unresolved, import/extensions */
 import InfiniteList, { Item } from 'terra-infinite-list/lib/index';
 import Placeholder from 'terra-doc-template/lib/Placeholder';
+import classNames from 'classnames/bind';
 /* eslint-enable import/no-unresolved, import/extensions */
 
 import MyProgressiveLoadingIndicator from '../common/MyProgressiveLoadingIndicator';
 import MyExampleContainer from '../common/MyExampleContainer';
+import styles from './InfiniteListDocExampleCommon.scss';
+
+const cx = classNames.bind(styles);
 
 const ProgressiveLoadingExample = () => (
   <MyExampleContainer>
@@ -14,7 +18,7 @@ const ProgressiveLoadingExample = () => (
       progressiveLoadingIndicator={<MyProgressiveLoadingIndicator />}
     >
       <Item key="item-1">
-        <Placeholder title="Item 1" style={{ height: '50px' }} />
+        <Placeholder title="Item 1" className={cx('placeholder')} />
       </Item>
     </InfiniteList>
   </MyExampleContainer>