+++ packages/terra-layout/tests/wdio/layout-slide-panel-spec.js
@@ -1,7 +1,6 @@
 /* globals $ */
-describe('Toggle the slide panel and hidden styles', () => {
-  before(() => browser.setViewportSize(Terra.viewports('medium')[0]));
-  beforeEach(() => {
+Terra.describeViewports('Toggle the slide panel and hidden styles', ['medium'], () => {
+  before(() => {
     browser.url('/#/raw/tests/terra-layout/layout/layout-slide-panel-example');
     browser.waitForExist('[class*=_panel_][aria-hidden="true"]');
   });
@@ -10,8 +9,7 @@ describe('Toggle the slide panel and hidden styles', () => {
 });
 
 describe('Toggle the slide panel click', () => {
-  before(() => browser.setViewportSize(Terra.viewports('medium')[0]));
-  beforeEach(() => {
+  before(() => {
     browser.url('/#/raw/tests/terra-layout/layout/layout-slide-panel-example');
     browser.click('#test-toggle-1');
     browser.waitForExist('[class*=_panel_][aria-hidden="false"]');
@@ -21,8 +19,7 @@ describe('Toggle the slide panel click', () => {
 });
 
 describe('Toggle the slide panel double click', () => {
-  before(() => browser.setViewportSize(Terra.viewports('medium')[0]));
-  beforeEach(() => {
+  before(() => {
     const toggleBtn1 = $('#test-toggle-1');
     browser.pause(150);
     const toggleBtn2 = $('#test-toggle-2');