+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -8,6 +8,7 @@ Unreleased
 
 ### Changed
 * Removed all inline styles from Dev-site components and implemented the same styles using external css.
+* Update wdio spec to use Terra.describeViewports
 
 3.2.0 - (June 12, 2019)
 ------------------